MicrocontrollerServer.Microcontroller (MicrocontrollerServer v0.1.0)

The Microcontroller context.

change_device(device, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking device changes.

Examples

iex> change_device(device)
%Ecto.Changeset{data: %Device{}}

change_reading(reading, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking reading changes.

Examples

iex> change_reading(reading)
%Ecto.Changeset{data: %Reading{}}

change_sensor(sensor, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking sensor changes.

Examples

iex> change_sensor(sensor)
%Ecto.Changeset{data: %Sensor{}}

create_device(attrs \\ %{})

Creates a device.

Examples

iex> create_device(%{field: value})
{:ok, %Device{}}

iex> create_device(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_reading(attrs \\ %{})

Creates a reading.

Examples

iex> create_reading(%{field: value})
{:ok, %Reading{}}

iex> create_reading(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_readings(attrs \\ [])

@spec create_readings(attrs :: [map()]) ::
  {:ok, %{readings: {number_of_rows :: integer(), nil}}}
  | {:error, [Ecto.Changeset.t()]}

Insert multiple readings at the same time. Similar to create_reading/1, only that it retuns an array when it's completed.

create_sensor(attrs \\ %{})

Creates a sensor.

Examples

iex> create_sensor(%{field: value})
{:ok, %Sensor{}}

iex> create_sensor(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_device(device)

Deletes a device.

Examples

iex> delete_device(device)
{:ok, %Device{}}

iex> delete_device(device)
{:error, %Ecto.Changeset{}}

delete_reading(reading)

Deletes a reading.

Examples

iex> delete_reading(reading)
{:ok, %Reading{}}

iex> delete_reading(reading)
{:error, %Ecto.Changeset{}}

delete_sensor(sensor)

Deletes a sensor.

Examples

iex> delete_sensor(sensor)
{:ok, %Sensor{}}

iex> delete_sensor(sensor)
{:error, %Ecto.Changeset{}}

get_device!(id)

Gets a single device.

Raises Ecto.NoResultsError if the Device does not exist.

Examples

iex> get_device!(123)
%Device{}

iex> get_device!(456)
** (Ecto.NoResultsError)

get_device_by_fields(controller_id, user_id, location_id)

Gets an existing device by controller_id.

Examples

iex> get_device_by_fields(1, 2, 3)
{:ok, %Device{controller_id: 1, user_id: 2, location_id: 3}}

iex> get_device_by_fields(bad_value, bad_value, bad_value)
nil

get_reading!(id)

Gets a single reading.

Raises Ecto.NoResultsError if the Reading does not exist.

Examples

iex> get_reading!(123)
%Reading{}

iex> get_reading!(456)
** (Ecto.NoResultsError)

get_sensor!(id)

Gets a single sensor.

Raises Ecto.NoResultsError if the Sensor does not exist.

Examples

iex> get_sensor!(123)
%Sensor{}

iex> get_sensor!(456)
** (Ecto.NoResultsError)

list_devices()

Returns the list of devices.

Examples

iex> list_devices()
[%Device{}, ...]

list_readings()

Returns the list of readings.

Examples

iex> list_readings()
[%Reading{}, ...]

list_sensors()

Returns the list of sensors.

Examples

iex> list_sensors()
[%Sensor{}, ...]

list_sensors_for_device(device_id)

@spec list_sensors_for_device(device_id :: integer()) :: [
  MicrocontrollerServer.Microcontroller.Sensor.t()
]

Gets all of the sensor values for the given device.

Examples

iex> list_sensors_for_device(device_id)
[%Sensor{}, %Sensor{}]

load_sensor_with_device(sensors)

Load the ecto relation with the device for a sensor.

load_sensor_with_readings(sensors)

Load the ecto relation with readings for a sensor.

update_device(device, attrs)

Updates a device.

Examples

iex> update_device(device, %{field: new_value})
{:ok, %Device{}}

iex> update_device(device, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_reading(reading, attrs)

Updates a reading.

Examples

iex> update_reading(reading, %{field: new_value})
{:ok, %Reading{}}

iex> update_reading(reading, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_sensor(sensor, attrs)

Updates a sensor.

Examples

iex> update_sensor(sensor, %{field: new_value})
{:ok, %Sensor{}}

iex> update_sensor(sensor, %{field: bad_value})
{:error, %Ecto.Changeset{}}