MicrocontrollerServer.Services.AuthServices.MicrocontrollerAuthService (MicrocontrollerServer v0.1.0)

Service file for perfoming requests to the microcontroller authentication server.

authenticate_token(token)

@spec authenticate_token(token :: binary()) ::
  {:ok,
   %{user_id: integer(), location_id: integer(), contrroller_id: integer()}}
  | {:error, :authentication_failed}

Authenticate the token against the authentication servers. If the token is okay then the authentication server witll give basic controller infromation like the Controller ID, user ID and location ID (the location ID is used to send information to the controlers en-masse).

Examples

iex> authenticate_token("CORRECT_TOKEN")
{:ok, %{user_id: 1, location_id: 2, controller_id: 3}}

iex> authenticate_token("INVALID_TOKEN")
{:error, :authentication_failed}

iex> authenticate_token("INACCESSIBLE_SERVER")
{:error, :authentication_failed}