MicrocontrollerServerWeb.MicrocontrollerSocket (MicrocontrollerServer v0.1.0)

Socket for microcontrollers trying to authenticate to the server. This file handles the incoming connections, checks whether the API token which they are sending are valid and in case they are it allows the request to be handled by the correct API version of the channel.

When the client connects to the socket a message is sent to it containing the required metadata.

api_token_regex(api_token)

@spec api_token_regex(api_token :: String.t()) ::
  {:error, :invalid_token} | {:ok, String.t()}

Check the API token against a preset regex rule. The rule is that the API token must start with API_TOKEN_MC_ and is then followed by at least 16 characters which can be uppercase and lowercaser letters alongside numbers.

Examples

iex> api_token_regex("API_TOKEN_MC_VZGkp2vvJJjHj3qZ")
{:ok, "API_TOKEN_MC_VZGkp2vvJJjHj3qZ"}

iex> api_token_regex("API_TOKEN_MC_VZGkp2vvJ")
{:error, :invalid_token}

iex> api_token_regex("oAPI_TOKEN_MC_VZGkp2vvJ")
{:error, :invalid_token}

iex> api_token_regex("oAPI_TOKEN_MC_VZGkp2vvJ@")
{:error, :invalid_token}

iex> api_token_regex("API_TOKEN_MC_VZGkp2vvJJjHj3q@")
{:error, :invalid_token}

iex> api_token_regex(nil)
{:error, :invalid_token}

authenticate_token(api_token)

@spec authenticate_token(any()) ::
  {:ok, integer(), integer(), integer()} | {:error, :failed_authentication}

Authenticate the device with the microcontroller authentication server.

The function returns the following if successful:

:ok

user_id

location_id

controller_id

Examples

iex> authenticate_token("API_TOKEN_MC_VZGkp2vvJJjHj3qZ")
{:ok, 1, 2, 3}

iex> authenticate_token("API_TOKEN_MC_INVALIDTOKENTEST")
{:error, :failed_authentication}

extract_api_key(arg1)

@spec extract_api_key(any() | {String.t(), String.t()}) :: String.t() | nil

Retrieve the x-api-key header from a list of x_headers. These headers are present in the connect_info variable in the connect/3 function call.

Examples

iex> extract_api_key({"x-api-key", "API_TOKEN_MC_VZGkp2vvJJjHj3qZ"})
"API_TOKEN_MC_VZGkp2vvJJjHj3qZ"

iex> extract_api_key({"x-api-key1", "API_TOKEN_MC_VZGkp2vvJJjHj3qZ"})
nil

generate_metadata_message(socket)

@spec generate_metadata_message(socket :: Phoenix.Socket.t()) :: {:ok, map()} | :error

Generate the metadata message for sockets. This is a message which gives the socket basic information about itself to be used in future API calls.

Examples

iex> generate_metadata_message(%Phoenix.Socket{assigns: %{device: %MicrocontrollerServer.Microcontroller.Device{location_id: 1, user_id: 2, controller_id: 3}}})
{:ok, %{ref: nil, payload: %{controller_id: 3, user_id: 2, location_id: 1}, topic: nil, join_ref: nil, event: "metadata"}}

iex> generate_metadata_message(%Phoenix.Socket{})
:error